@@ -1,11 +1,12 @@
-import React from 'react';
-import { useSelector } from 'react-redux';
+import React, {useContext} from 'react';
 
 import ProductItem from '../components/Products/ProductItem';
+import { ProductContext } from '../context/product-context';
 import './Products.css';
 
 const Products = props => {
-  const productList = useSelector(state => state.shop.products);
+  const productList = useContext(ProductContext).products;
+  
   return (
     <ul className="products-list">
       {productList.map(prod => (
